% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMAauto.r
\name{ARMAauto}
\alias{ARMAauto}
\title{compute the autocovariance function of an ARMA process}
\usage{
ARMAauto(ar = NULL, ma = NULL, lag.max)
}
\arguments{
\item{ar}{numeric vector of AR coefficients}

\item{ma}{numeric vector of MA coefficients}

\item{lag.max}{Largest autocovariance lag required}
}
\value{
autocovariances at lags 0 through lag.max
}
\description{
Background: function computes autocovariances of ARMA (p,q) from lag zero
	to lag.max, with inputs ar and ma.  Format:
	(1 - ar[1]z ... - ar[p]z^p) X_t = (1 + ma[1]z ...+ ma[q]z^q) WN
          For absent AR or MA portions, pass in NULL
}
