% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_VARMA}
\alias{auto_VARMA}
\title{computes autocovariances of SVARMA usng frequency domain}
\usage{
auto_VARMA(param, p, q, ps, qs, season, grid, maxlag)
}
\arguments{
\item{param}{- matrix of dimension m x (p+q+ps+qs+1), equals [ phi | theta | phiseas | thetaseas | sigma ]
  phi: block matrix of dimension N x N*p of VAR coefficients
  theta: block matrix of dimension N x N*q of VMA coefficients
phiseas:  block matrix of dimension N x N*ps of SVAR coefficients
thetaseas:  block matrix of dimension N x N*qs of SVMA coefficients
  sigma: N x N covariance matrix of white noise}

\item{p}{- AR order}

\item{q}{- MA order}

\item{ps}{- seasonal AR order}

\item{qs}{- seasonal MA order}

\item{season}{- period (e.g. monthly = 12)}

\item{grid}{- Riemann mesh size}

\item{maxlag}{- maximum autocovariance lag needed}
}
\value{
autocovariances at lags 0 through maxlag, as array of dimension m x m x (maxlag+1)
}
\description{
Backgroud:
  function computes autocovariances of SVARMA (p,q,ps,qs) from lag zero
	to maxlag, with array inputs phi and theta.  SVARMA equation:
  (1 - phi[1]B ... - phi[p]B^p) (1 - Phi[1]B^s ... - Phi[ps]B^{s*ps}) X_t =
  (1 + theta[1]B ... + theta[q]B^q) (1 + Theta[1]B^s ... + Theta[qs]B^{s*qs}) WN_t.
}
