% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bfs}
\alias{bfs}
\title{Business Formation Statistics Weekly Data}
\format{
## `bfs`
A data frame with 757 rows and 10 columns:
\describe{
  \item{Year}{Four digit year}
  \item{Week}{Week of the year based on IRS processing. Please
              see below for how weeks match to calendar
              dates. Please note that some years (2008, 2014,
              and 2020) have 53 weeks.}
 \item{BA_NSA}{Not seasonally adjusted Business Application national series}
 \item{HBA_BSA}{Not seasonally adjusted High-Propensity Business
Application national series}
 \item{WBA_NSA}{Not seasonally adjusted Businesses Applications
with Planned Wages national series}
 \item{CBA_NSA}{Not seasonally adjusted Business Applications
from Corporations national series}
 \item{YY_BA_NSA}{Calculated year-to-year percentage changes for
same week a year ago for the Business
Application series. No values are available for first
year of the series (2006) or for week 53s.}
 \item{YY_HBA_NSA}{Calculated year-to-year percentage changes for
same week a year ago for the High-Propensity
Business Application series. No values are
available for first year of the series (2006) or for
week 53s.}
 \item{YY_WBA_NSA}{Calculated year-to-year percentage changes for
same week a year ago for the Businesses
Applications with Planned Wages series. No
values are available for first year of the series
(2006) or for week 53s.}
 \item{YY_CBA_NSA}{Calculated year-to-year percentage changes for
same week a year ago for the Business
Applications from Corporations series. No values
are available for first year of the series (2006) or
for week 53s.}
}
}
\source{
<https://www.census.gov/data/experimental-data-products/weekly-business-formation-statistics.html>
}
\usage{
bfs
}
\description{
BFS measure business initiation activity and the cycle from initiation
to realized business formation. BFS gives an early look at business
formation activity within the U.S..
}
\keyword{datasets}
