% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2day.r
\name{date2day}
\alias{date2day}
\title{converts month-day-year date, and returns day index}
\usage{
date2day(month, day, year)
}
\arguments{
\item{month}{1 through 12, corresponding to the date's month}

\item{day}{1 through 31, corresponding to the date's day}

\item{year}{0 through ???, corresponding to the date's year (A.D.)}
}
\value{
day.count: day index within a year of the given date, from 1 to 366
}
\description{
converts month-day-year date, and returns day index
}
