% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datefinder.r
\name{datefinder}
\alias{datefinder}
\title{converts a date in month-day-year format and finds the day of week}
\usage{
datefinder(dayofweek, week, month, year)
}
\arguments{
\item{date}{a given date in month-day-year format,
a 3-element vector}
}
\value{
dayofweek: number 1 through 7, where 1 is Sunday, 2 is Monday, etc.
}
\description{
converts a date in month-day-year format and finds the day of week
}
\note{
calibrated according to Jan 1, 1600 being a Saturday
}
