% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2date.r
\name{day2date}
\alias{day2date}
\title{converts a day index to a date in month-day-year format}
\usage{
day2date(day, start.date)
}
\arguments{
\item{day}{integer, the number of days past the given start.date
for which the date is desired; negative values reckon
into the past}

\item{start.date}{a given starting date in month-day-year format,
a 3-element vector}
}
\value{
end.date: the ending date  in month-day-year format,
		a 3-element vector
}
\description{
converts a day index to a date in month-day-year format
}
