% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_monthComp.R
\name{gen_monthComp}
\alias{gen_monthComp}
\title{Generate multivariate seasonal monthly component}
\usage{
gen_monthComp(n, Phi, Ndim, Sig, burn = 1000)
}
\arguments{
\item{n}{length of series}

\item{Phi}{seasonal autoregressive parameter}

\item{Ndim}{dimensionality of each observation}

\item{Sig}{covariance martrix white noise component}

\item{burn}{burn in (defaults to 1000)}
}
\value{
Ndim x n matrix of observations
}
\description{
generates a component that is monthly (12) difference stationary
}
