% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trendComp.R
\name{gen_trendComp}
\alias{gen_trendComp}
\title{Generate multivariate trend component}
\usage{
gen_trendComp(n, Ndim, Sig, burn = 1000)
}
\arguments{
\item{n}{length of series}

\item{Ndim}{dimensionality of each observation}

\item{Sig}{covariance martrix white noise component}

\item{burn}{burn in (defaults to 1000)}
}
\value{
Ndim x n matrix of observations
}
\description{
generates a component that is first difference stationary
}
