% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEigenValues}
\alias{getEigenValues}
\title{Eigenvalue decompostion}
\usage{
getEigenValues(M)
}
\arguments{
\item{X}{- square matrix}
}
\value{
The eigenvalues and corresponding right eigenvectors are stored in
    eigval and eigvec, respectively
}
\description{
Wrapper function for armadillo `eig_gen`.
Eigen decomposition of dense general (non-symmetric/non-hermitian) square matrix X
}
