% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGCD.r
\name{getGCD}
\alias{getGCD}
\title{computes the generalized Cholesky decomposition}
\usage{
getGCD(Sigma, Rank)
}
\arguments{
\item{Sigma}{symmetric, non-negative definite matrix}

\item{Rank}{presumed rank of Sigmma, less than or equal
to the dimension of the matrix}
}
\value{
list of length 2 giving L.mat and D.mat.
   L.mat: rectangular, lower Cholesky factor
	D.mat: vector of diagonal entries in Cholesky decomposition
}
\description{
computes the generalized Cholesky decomposition
}
