% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethol.r
\name{gethol}
\alias{gethol}
\title{Generates holiday regressors from holiday dates}
\usage{
gethol(hol.dates, hol.fore, hol.aft, start.date, end.date)
}
\arguments{
\item{hol.dates}{List of dates in month-day-year format;
this is a matrix with one row for each holiday date,
and 3 columns for month, day in month, and year}

\item{hol.fore}{Number of days before holiday, where effect begins}

\item{hol.aft}{Number of days after holiday, where effect ends}

\item{start.date}{Start date of time series in month-day-year format,
a 3-element vector}

\item{end.date}{End date of time series in month-day-year format,
a 3-element vector}
}
\value{
subset.hol.reg: regressors corresponding to the holiday effect,
		with an activity window determined by hol.fore and hol.aft,
		and at dates between start.date and end.date
}
\description{
Notes:  this version presumes start.date and end.date are in separate years
}
