% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvar.forecast.r
\name{mvar.forecast}
\alias{mvar.forecast}
\title{Compute multi-step forecasts and predictors of a multivariate process
	via Levinson-Durbin algorithm}
\usage{
mvar.forecast(x.acf, z, needMSE)
}
\arguments{
\item{x.acf}{Array of dimension N x T x N of autocovariances for process w_t,
where there are N series, of total length T each.}

\item{z}{Differenced data as N x (T+H) matrix, with missing values at
various time points.  Presumes first T observations are not missing,
and latter H observations are missing, being encoded
with 1i in that entry.  That is,
Im(z[,t]) = rep(1i,N) encodes missing values.}

\item{needMSE}{A binary flag, with 1 indicating that forecast MSE should be computed;
else with a 0 this will be skipped (runs faster in this mode).}
}
\value{
list containing preds.x and pred.stack
	preds.x: N x (T+H) matrix of data with forecasts, where H
		is the total number of time indices with missing values.
	pred.stack: NT x H matrix of predictors, which can be used to
		obtain uncertainty.
}
\description{
Compute multi-step forecasts and predictors of a multivariate process
	via Levinson-Durbin algorithm
}
