% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polymul_mat}
\alias{polymul_mat}
\title{polumul_mat - Documentation update needed}
\usage{
polymul_mat(amat, bmat)
}
\value{
product amat(z) * bmat(z)
}
\description{
polumul_mat - Documentation update needed
}
