% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymulMat.r
\name{polymulMat}
\alias{polymulMat}
\title{compute the product of two matrix polynomials}
\usage{
polymulMat(amat, bmat)
}
\arguments{
\item{amat}{array of polynomial coefficients, where amat[,,1] is the zeroth coefficient,
amat[,,2] is the first coefficient, etc.}

\item{bmat}{vector of polynomial coefficients, where bmat[,,1] is the zeroth coefficient,
bmat[,,2] is the first coefficient, etc.}
}
\value{
array of polynomial coefficients for cmat(z) = mata(z) * matb(z),
		where cmat[,,1] is the zeroth coefficient,
		      cmat[,,2] is the first coefficient, etc.
}
\description{
compute the product of two matrix polynomials
}
