% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysum.r
\name{polysum}
\alias{polysum}
\title{compute the sum of two polynomials,}
\usage{
polysum(a, b)
}
\arguments{
\item{a}{symmetric vector of polynomial coefficients,
where a[1] is the zeroth coefficient,
a[2] is the first coefficient, etc.}

\item{b}{symmetric vector of polynomial coefficients,
where b[1] is the zeroth coefficient,
b[2] is the first coefficient, etc.}
}
\value{
symmetric vector of polynomial coefficients for c(z) = a(z) + b(z),
		where c[1] is the zeroth coefficient, c[2] is the first coefficient, etc.
}
\description{
compute the sum of two polynomials,
}
