% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.add.r
\name{sigex.add}
\alias{sigex.add}
\title{Build the model by adding on another latent component}
\usage{
sigex.add(mdl, vrank, class, order, bounds, name, delta)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{vrank}{Vector of integers between 1 and N, corresponding
to indices of non-zero Schur complements in the GCD
of the innovations' covariance matrix for the new latent component}

\item{class}{Character string of t.s. model type for the new latent component}

\item{order}{Vector of model order bounds:
four numbers, gives bounds for rho and omega,
   the cycle parameters of the new latent component
   rho lies in (bounds[1],bounds[2])
   omega lies in (bounds[3],bounds[4])}

\item{name}{Character string giving the latent component's name}

\item{delta}{Differencing polynomial (corresponds to delta(B) in Background)
written in format c(delta0,delta1,...,deltad)}
}
\value{
mdl: the updated sigex model, a list object
}
\description{
Build the model by adding on another latent component
}
