% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.blocktoep.r
\name{sigex.blocktoep}
\alias{sigex.blocktoep}
\title{Generates lower triangular array for block Toeplitz matrix}
\usage{
sigex.blocktoep(x.array)
}
\arguments{
\item{x.array}{Array of dimension N x N x H, where x.array[,,1] is gamma(0)}
}
\value{
x.toep: array of dimension N x H x N x H, where x.toep[,j,,k]
}
\description{
Generates lower triangular array for block Toeplitz matrix
}
