% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.bundle.r
\name{sigex.bundle}
\alias{sigex.bundle}
\title{Bundles into a list object several features}
\usage{
sigex.bundle(data.ts, transform, mdl, psi)
}
\arguments{
\item{data.ts}{A T x N matrix ts object,
corresponding to N time series of length T}

\item{transform}{A character indicating an instantaneous
transformation to be applied; current options are
"none", "log", and "logistic"}

\item{mdl}{The specified sigex model, a list object}

\item{psi}{A vector of all the real hyper-parameters}
}
\value{
analysis: a list object of the inputs
}
\description{
Background:
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter	manifold)
}
