% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.cast.r
\name{sigex.cast}
\alias{sigex.cast}
\title{Computes forecasts and aftcasts, without uncertainty;
	a faster version of sigex.midcast, if you don't need midcasts}
\usage{
sigex.cast(psi, mdl, data.ts, leads)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}

\item{leads}{An integer sequence of desired casts; include index t
to obtain an estimate of x_t.  These integers don't have
to be a subset of {1,2,...,T}.  Include integers greater than
T to get forecasts, or less than 1 to get aftcasts.}
}
\value{
x.casted: N x (T+H) matrix of aftcasts, data, and forecasts, where H
		is the total number of forecasts and aftcasts
}
\description{
Computes forecasts and aftcasts, without uncertainty;
	a faster version of sigex.midcast, if you don't need midcasts
}
