% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.conditions.r
\name{sigex.conditions}
\alias{sigex.conditions}
\title{Computes condition number for a covariance matrix}
\usage{
sigex.conditions(data.ts, psi, mdl)
}
\arguments{
\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}

\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}
}
\value{
conds: a S x N matrix of condition numbers, where
		S is the number of components.  Each row gives
		the N condition numbers for the innovation
		covariance matrix of the corresponding latent component.
}
\description{
Computes condition number for a covariance matrix
}
