% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.constrainreg.r
\name{sigex.constrainreg}
\alias{sigex.constrainreg}
\title{Determines constraint matrix for regressors}
\usage{
sigex.constrainreg(mdl, data.ts, regindex, combos)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}

\item{regindex}{A list object of N elements, containing indices J_i
for i=1,...,N of the regressors for the ith series involved.}

\item{combos}{A vector of linear combinations, the subvectors have length
J_i corresponding to regindex, and a first element which is Q
BUT: if set to NULL, then all regressors in regindex are constrained
  to have the same value.}
}
\value{
constraint: row vector of the form [Q , C], with C (constraint.mat)
    the vector of constraints and Q (constraint.vec) the
    constraint constant, such that C psi = Q.
    When combos=NULL, constraint will be a matrix
}
\description{
Determines constraint matrix for regressors
}
