% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.daily2weekly.r
\name{sigex.daily2weekly}
\alias{sigex.daily2weekly}
\title{Embeds a daily time series as a weekly time series}
\usage{
sigex.daily2weekly(data.ts, first.day, start.date)
}
\arguments{
\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}

\item{first.day}{A number between 1 and 7 that indicates what day of the week
should correspond to the first component of the 7-vector.
So 1 = Sunday, 2 = Monday, 3 = Tuesday, 4 = Wednesday,
  5 = Thursday, 6 = Friday, 7 = Sunday.}

\item{start.date}{A given starting date in month-day-year format,
a 3-element vector}
}
\value{
data.ts: a T x 7 matrix ts object; any  values to be imputed
		are marked with NA in that entry.  The NA is for missing value.
}
\description{
Embeds a daily time series as a weekly time series
}
