% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.delta.r
\name{sigex.delta}
\alias{sigex.delta}
\title{Compute differencing polynomial with factors omitted}
\usage{
sigex.delta(mdl, omits)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{omits}{Indices of components that are to be omitted,
when computing the product of differencing polynomials
for the various components.}
}
\value{
Updated differencing polynomial,
	written in format c(delta0,delta1,...,deltad)
}
\description{
Compute differencing polynomial with factors omitted
}
