% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.eta2psi.r
\name{sigex.eta2psi}
\alias{sigex.eta2psi}
\title{Transform eta to psi}
\usage{
sigex.eta2psi(eta, constraint)
}
\arguments{
\item{eta}{free variables of real numbers.  Must have length given by
dim(constraint.mat)[2] - dim(constraint.mat)[1]}

\item{constraint}{Matrix of the form [Q , C], with C (constraint.mat)
the matrix of constraints and Q (constraint.vec) the vector
of constraint constants, such that C psi = Q.}
}
\value{
psi: a vector of all the real hyper-parameters.
}
\description{
Transform eta to psi
}
