% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.fixed.r
\name{sigex.fixed}
\alias{sigex.fixed}
\title{Computes all nontrend fixed regression effects}
\usage{
sigex.fixed(data.ts, mdl, series, param, type)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{mdl}{The specified sigex model, a list object}

\item{series}{Integer between 1 and N, the index of the individual series for
which regression effects are being computed.}

\item{param}{The model parameters entered into a list object}

\item{type}{A string designating the name of the regression effect}
}
\value{
mean.mat: length T time series consisting of regression effects.
		This has the format X %*% beta, where each column of
		X is a regressor of "type", and beta consists of the corresponding
		regression parameters.
}
\description{
Computes all nontrend fixed regression effects
}
