% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.frf.r
\name{sigex.frf}
\alias{sigex.frf}
\title{Computes signal extraction filter frequency response function}
\usage{
sigex.frf(data.ts, param, mdl, sigcomps, grid)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{The model parameters entered into a list object}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal}

\item{grid}{Desired number of frequencies for spectrum calculations}
}
\value{
frf.wk: array of dimension c(N,N,grid), with complex number entries
}
\description{
Computes signal extraction filter frequency response function
}
