% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.getcycle.r
\name{sigex.getcycle}
\alias{sigex.getcycle}
\title{Compute AR and MA polynomials for Butterworth and Balanced cycles}
\usage{
sigex.getcycle(cycle.order, rho, omega)
}
\arguments{
\item{cycle.order}{Integer order of the cycle}

\item{rho}{Persistency, a value between 0 and 1}

\item{omega}{Frequency in units of pi, a value between 0 and 1.
(The period is 2/omega)}
}
\value{
list object with cycle.AR and cycle.MA
	cycle.AR: full AR polynomial in format c(1,-phi1,...,-phip)
	cycle.MA: full MA polynomial in format c(1,theta1,...,thetaq)
}
\description{
Background:
	A stochastic cycle is a particular parametrization of an ARMA process,
	governed by its order, its persistency (rho), and its period (2/omega)
}
