% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.getfrf.r
\name{sigex.getfrf}
\alias{sigex.getfrf}
\title{Computes frequency response function for desired signal and plots it}
\usage{
sigex.getfrf(data.ts, param, mdl, sigcomps, plotit = TRUE, grid)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{The model parameters entered into a list object}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal}

\item{plotit}{Boolean flag for whether frf should be plotted;
only plots if N <= 3}

\item{grid}{Desired number of frequencies for spectrum calculations}
}
\value{
frf.comp:  array of dimension c(N,N,grid), with complex number entries
}
\description{
Computes frequency response function for desired signal and plots it
}
