% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.glr.r
\name{sigex.glr}
\alias{sigex.glr}
\title{Computes the difference of -2*log(Gaussian likelihood) for
	two models, the nested lik minus nesting lik}
\usage{
sigex.glr(data.ts, psi.nested, psi.nesting, mdl.nested, mdl.nesting)
}
\arguments{
\item{data.ts}{A T x N matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{psi.nested}{psi of the nested model}

\item{psi.nesting}{psi of the nesting model}

\item{mdl.nested}{The specified nested sigex model, a list object}

\item{mdl.nesting}{The specified nesting sigex model, a list object}
}
\value{
glr: difference of log likelihoods,
	dof: degrees of freedom, given by different in number of parameters
}
\description{
Computes the difference of -2*log(Gaussian likelihood) for
	two models, the nested lik minus nesting lik
}
