% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.graph.r
\name{sigex.graph}
\alias{sigex.graph}
\title{Adds signal extraction estimates, with shaded bands,
	to an existing time series plot}
\usage{
sigex.graph(
  extract,
  reg = NULL,
  start.date,
  period,
  series,
  displace,
  color,
  fade
)
}
\arguments{
\item{extract}{T x N matrix of the signal estimates, e.g. output of sigex.extract}

\item{reg}{A T x N matrix of fixed effects, to be added to extraction.
Note: only the column of reg corresponding to "series" is utilized.
(To incorporate multiple reg effects, add these all up before hand.)}

\item{start.date}{Date of first time obersvation; the
format is c(year,season)}

\item{period}{Number of seasons per year}

\item{series}{Index of the particular series under consideration}

\item{displace}{Gives a vertical shift to the plot}

\item{color}{Given as a number in the range of colors()}

\item{fade}{Gives shading proportion for uncertainty (NULL if none)}
}
\value{
NA
}
\description{
Adds signal extraction estimates, with shaded bands,
	to an existing time series plot
}
