% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.hi2low.r
\name{sigex.hi2low}
\alias{sigex.hi2low}
\title{Embeds a given high frequency filter as a low frequency filter
    for a vector embedding}
\usage{
sigex.hi2low(filter.hi, hi.freq, low.freq, shift.hi)
}
\arguments{
\item{filter.hi}{Given scalar high frequency filter of length L}

\item{hi.freq}{Sampling frequency of high frequency time series}

\item{low.freq}{Sampling frequency of low frequency time series}

\item{shift.hi}{The integer offset for the high frequency filter:
filter coefficients have indices -shift.hi,...,0,...,L-1-shift.hi
set shift.hi = 0 for a causal filter}
}
\value{
filter.low: array s x s x M,
    where s = hi.freq/low.freq (must be an integer)
    and M is length, which depends on s, L, and shift.hi
  shift.low: integer offset for the low frequency filter
}
\description{
Embeds a given high frequency filter as a low frequency filter
    for a vector embedding
}
