% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.i2rag.r
\name{sigex.i2rag}
\alias{sigex.i2rag}
\title{Reads in a multivariate time series and generates a list of missing value indices}
\usage{
sigex.i2rag(z)
}
\arguments{
\item{z}{raw data as N x T matrix with missing values at various time points.
    Missing values are at any 1 <= t <= T, and occur for some of the N series
(the ragged case), and are denoted with a 1i.  That is,
    Im(z[,t]) = rep(1i,N) or subset thereof encodes missing values.}
}
\value{
list containing leads.rag and ragged
	leads.rag: an integer sequence of indices that have at least one missing value
		These integers are a subset of {1,2,...,T}.
  ragged: a list with number of items equal to length of leads.rag
    within the sample {1,2,...,T}.  (ragged=NULL if there are none.)
    Each element contains indices of vector components that are missing.
    Note: ragged[[j]] must be non-empty, where leads[j] is the time
     where at least one missing value (in-sample) occurs.
}
\description{
Reads in a multivariate time series and generates a list of missing value indices
}
