% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.lik.r
\name{sigex.lik}
\alias{sigex.lik}
\title{Computes -2*log(Gaussian likelihood) of model}
\usage{
sigex.lik(psi, mdl, data.ts, debug = TRUE)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{debug}{Set to TRUE if lik values should be printed to screen}
}
\value{
Value of the divergence
}
\description{
Computes -2*log(Gaussian likelihood) of model
}
