% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.lpfiltering.r
\name{sigex.lpfiltering}
\alias{sigex.lpfiltering}
\title{Compute signal extraction estimates with uncertainty
	for trend and cycle, by combining WK filter for trend-cycle
	(specified by trendcyclecomp) with LP filter of cutoff.}
\usage{
sigex.lpfiltering(
  mdl,
  data.ts,
  trendcyclecomp,
  sigcomps,
  psi,
  cutoff,
  grid,
  window,
  trunc,
  trendFlag
)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object}

\item{trendcyclecomp}{The (single) index of the trend-cycle component}

\item{sigcomps}{Provides indices of a desired component that
    is disjoint from trend-cycle, so that MSEs of
trend+sigcomps and cycle+sigcomps are computed.
 (Pass in sigcomps = NULL to just get trend and cycle MSEs.)}

\item{psi}{A vector of all the real hyper-parameters}

\item{cutoff}{A number between 0 and pi, with all frequencies < cutoff preserved}

\item{grid}{Desired number of frequencies for spectrum calculations}

\item{window}{Maximum index of the filter coefficients}

\item{trunc}{Truncation index for LP filter}

\item{trendFlag}{Boolean flag, TRUE for trend+signal, else get cycle+signal}
}
\value{
list object with lp.signal, upp, and low
	lp.signal: T x N matrix of the signal estimates
	upp: as lp.signal, plus twice the standard error
	low: as lp.signal, minus twice the standard error
}
\description{
Background:
	A sigex model consists of process x = sum y, for
	stochastic components y.  Each component process y_t
	is either stationary or is reduced to stationarity by
	application of a differencing polynomial delta(B), i.e.
		w_t = delta(B) y_t   is stationary.
	We have a model for each w_t process, and can compute its
	autocovariance function (acf), and denote its autocovariance
	generating function (acgf) via gamma_w (B).
	The signal extraction filter for y_t is determined from
	this acgf and delta.  The error spectral density calculations
	are found in:
	"Casting Vector Time Series: Algorithms for Forecasting,
	Imputation, and Signal Extraction," McElroy (2018).
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter	manifold)
}
\details{
Notes:
	Starts with LP an ideal low-pass filter,
	and applies LP*WK filter with cutoff parameter  to
	each component, where the filter has been truncated to
	length 2*window + 1.  The output will be LP*WK filter applied to data,
  	which is forecast and aftcast extended by window units, covering time points
  	1-window ..., T+window.  This gives trend and cycle at times 1,...,T.
      Take grid >> window, else numerical issues arise
}
