% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.lpwk.r
\name{sigex.lpwk}
\alias{sigex.lpwk}
\title{Computes signal extraction filter coefficients
	for trend and cycle, by combining WK filter for trend-cycle
	(specified by trendcyclecomp) with LP filter of cutoff.}
\usage{
sigex.lpwk(data.ts, param, mdl, trendcyclecomp, grid, len, cutoff, trunc)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}

\item{trendcyclecomp}{The (single) index of the trend-cycle component}

\item{grid}{Desired number of frequencies for spectrum calculations}

\item{len}{Max index of the filter coefficients}

\item{cutoff}{A number between 0 and pi, with all frequencies < cutoff preserved}

\item{trunc}{Truncation index for LP filter}
}
\value{
list object with psi.lpwk and psi.ilpwk
	psi.lpwk: array of dimension c(N,N,2*trunc+1), filter coefficients for trend
	psi.ilpwk: array of dimension c(N,N,2*trunc+1), filter coefficients for cycle
}
\description{
Background:
	A sigex model consists of process x = sum y, for
	stochastic components y.  Each component process y_t
	is either stationary or is reduced to stationarity by
	application of a differencing polynomial delta(B), i.e.
		w_t = delta(B) y_t   is stationary.
	We have a model for each w_t process, and can compute its
	autocovariance function (acf), and denote its autocovariance
	generating function (acgf) via gamma_w (B).
	The signal extraction filter for y_t is determined from
	this acgf and delta.
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter	manifold)
}
\details{
Notes: take grid >> len, else numerical issues arise
}
