% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.meaninit.r
\name{sigex.meaninit}
\alias{sigex.meaninit}
\title{Adds trend regressors to an existing model}
\usage{
sigex.meaninit(mdl, data.ts, d)
}
\arguments{
\item{mdl}{The specified sigex model, a list object.
  mdl[[1]] is mdlK, gives ranks of white noise covariance matrix
  mdl[[2]] is mdlType, a list giving t.s. model class, order, and bounds
  mdl[[3]] is mdlDiff, gives delta differencing polynomials
mdl[[4]] is list of regressors by individual series}

\item{data.ts}{A T x N matrix ts object}

\item{d}{Order of time polynomial trend regressor desired, labeled as "Trend".
 However, if a trend component exists in the model,
then this d is ignored, and the number of unit roots
is instead used to determine d.}
}
\value{
mdl: the updated sigex model, a list object
}
\description{
Adds trend regressors to an existing model
}
