% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.midcast.r
\name{sigex.midcast}
\alias{sigex.midcast}
\title{Computes predictors for variables at various indices}
\usage{
sigex.midcast(psi, mdl, data.ts, castspan)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{castspan}{A non-negative integer horizon giving number of fore- and aft-casts}
}
\value{
list containing casts.x and casts.var
	casts.x: N x H matrix of forecasts, midcasts, aftcasts, where H
		is the total number of time indices with missing values,
		given by cardinality( leads setminus {1,2,...,T} )
	casts.var: NH x NH matrix of covariances of casting errors.
		note that casts.var.array <- array(casts.var,c(N,H,N,H))
		corresponds to cast.var.array[,j,,k] equal to the
		covariance between the jth and kth casting errors
}
\description{
Computes predictors for variables at various indices
}
