% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.mlefit.r
\name{sigex.mlefit}
\alias{sigex.mlefit}
\title{Fit model to the data using ML estimation}
\usage{
sigex.mlefit(
  data.ts,
  param,
  constraint,
  mdl,
  method,
  thresh = Inf,
  hess = TRUE,
  whittle = FALSE,
  debug = FALSE,
  maxiter = 100
)
}
\arguments{
\item{data.ts}{A T x N matrix ts object; any missing values
must be encoded with 1i in that entry}

\item{param}{model parameters entered into
 a list object with an intuitive structure.
This is an initial specification to
     start the nonlinear optimization routines}

\item{constraint}{Matrix of the form [Q , C], with C (constraint.mat)
the matrix of constraints and Q (constraint.vec) the vector
of constraint constants, such that C psi = Q.
Use NULL if there are no constraints}

\item{mdl}{The specified sigex model, a list object}

\item{method}{"bfgs" for BFGS, "sann" for simulated annealing,
"cg" for conjugate gradient}

\item{thresh}{Pre-parameters theta satisfy |theta|< thresh;
set thresh = Inf if no thresholding is desired}

\item{hess}{A Boolean flag; if true, for BFGS it runs
another round of BFGS to get Hessian matrix}

\item{whittle}{A Boolean flag; if true, uses Whittle likelihood instead of
default Gaussian likelihood}

\item{debug}{A Boolean flag; if true, sets the DEBUGGING mode, which prints
psi.last and psi.now to the global field.  Also lik will be printed.
psi.now gives the parameter that crashed the likelihood (if it crashed),
psi.last gives the last good parameter that did not crash the lik.}

\item{maxiter}{Maximum number of iterations allowed in optimization routines}
}
\value{
list with mle and par.est
	mle: an object of type outputted by optim
	par.est: type param, with the estimated parameters filled in
}
\description{
Fit model to the data using ML estimation
}
