% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.momfit.r
\name{sigex.momfit}
\alias{sigex.momfit}
\title{Computes initial parameter estimates by method of moments}
\usage{
sigex.momfit(data.ts, param, mdl)
}
\arguments{
\item{data.ts}{A T x N matrix ts object; any missing values
must be encoded with 1i in that entry}

\item{param}{model parameters entered into
 a list object with an intuitive structure.
This is an initial specification to
     start the nonlinear optimization routines}

\item{mdl}{The specified sigex model, a list object}
}
\value{
par.new: type param, with the estimated covariance parameters filled in
}
\description{
Computes initial parameter estimates by method of moments
}
