% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.mvar2uvar.r
\name{sigex.mvar2uvar}
\alias{sigex.mvar2uvar}
\title{Transform multivariate parameters to implied univariate form}
\usage{
sigex.mvar2uvar(data.ts, param, mdl)
}
\arguments{
\item{data.ts}{A T x N matrix ts object; any missing values
must be encoded with 1i in that entry}

\item{param}{Model parameters entered into
 a list object with an intuitive structure.
This is an initial specification to
     start the nonlinear optimization routines}

\item{mdl}{The specified sigex model, a list object}
}
\value{
list with mdl.uni and univ.param
  	mdl.uni: output univariate model
	univ.param: param for univariate model
}
\description{
Transform multivariate parameters to implied univariate form
}
