% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.par2psi.r
\name{sigex.par2psi}
\alias{sigex.par2psi}
\title{Transform param to psi}
\usage{
sigex.par2psi(param, mdl)
}
\arguments{
\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}
}
\value{
psi: A vector of all the real hyper-parameters
}
\description{
Transform param to psi
}
