% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.par2zeta.r
\name{sigex.par2zeta}
\alias{sigex.par2zeta}
\title{Transform param to zeta}
\usage{
sigex.par2zeta(mdlPar, mdlType)
}
\arguments{
\item{mdlPar}{This is the portion of param
corresponding to mdlType}

\item{mdlType}{This is a component of mdl (the specified sigex model),
cited as mdl[[2]]}
}
\value{
zeta: portion of psi pre-parameter vector that corresponds to
     all hyper-parameters for t.s. models
}
\description{
Transform param to zeta
}
