% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.param2gcd.r
\name{sigex.param2gcd}
\alias{sigex.param2gcd}
\title{Utility that takes a real vector and inserts as
  	entries of a unit lower triangular matrix}
\usage{
sigex.param2gcd(L.psi, N, vrank)
}
\arguments{
\item{L.psi}{A vector of reals of length <= N(N-1)/2}

\item{N}{Dimension of the time series}

\item{vrank}{Vector of integers between 1 and N, corresponding
to indices of non-zero Schur complements in the GCD}
}
\value{
L.mat: unit lower	triangular matrix of dimension N x length(vrank)
	with entries given by L.psi
}
\description{
Utility that takes a real vector and inserts as
  	entries of a unit lower triangular matrix
}
