% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.portmanteau.r
\name{sigex.portmanteau}
\alias{sigex.portmanteau}
\title{Computes the portmanteau statistic for residuals}
\usage{
sigex.portmanteau(resids, lag, nump)
}
\arguments{
\item{resids}{A T x N matrix of residuals}

\item{lag}{Number of autocorrelation lags used in the statistic}

\item{nump}{Number of parameters that are estimated}
}
\value{
port and pval
 	port: value of the portmanteau test statistic
	pval: p-value of port, based on chi square with nump
		degrees of freedom
}
\description{
Computes the portmanteau statistic for residuals
}
