% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.precision.r
\name{sigex.precision}
\alias{sigex.precision}
\title{Compares signal extraction MSE arising from
 	multivariate fit with the implied univariate fit.}
\usage{
sigex.precision(data.ts, param, mdl, sigcomps)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{Model parameters entered into
 a list object with an intuitive structure.
This is an initial specification to
     start the nonlinear optimization routines}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal}
}
\value{
ratio of mse to mse.uni, the signal extraction MSE for
		multivariate and univariate models
}
\description{
Compares signal extraction MSE arising from
 	multivariate fit with the implied univariate fit.
}
