% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.psi2eta.r
\name{sigex.psi2eta}
\alias{sigex.psi2eta}
\title{Transform psi to eta}
\usage{
sigex.psi2eta(psi, constraint)
}
\arguments{
\item{psi}{A vector of real numbers containing all    hyper-parameters}

\item{constraint}{Matrix of the form [Q , C], with C (constraint.mat)
the matrix of constraints and Q (constraint.vec) the vector
of constraint constants, such that C psi = Q.}
}
\value{
list of nu and eta
    psi is a permuted vector of eta and nu, where nu are fixed variables,
    and eta consists of free variables
}
\description{
Transform psi to eta
}
