% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.psi2par.r
\name{sigex.psi2par}
\alias{sigex.psi2par}
\title{Transform psi to param}
\usage{
sigex.psi2par(psi, mdl, data.ts)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}
}
\value{
param:  model parameters entered into
	a list object with an intuitive structure.
}
\description{
Transform psi to param
}
