% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.rag2i.r
\name{sigex.rag2i}
\alias{sigex.rag2i}
\title{Read multivariate time series and insert 1i for indicated missing values}
\usage{
sigex.rag2i(leads.rag, ragged, z)
}
\arguments{
\item{leads.rag}{An integer sequence of indices that have at least one missing value.
These integers are a subset of {1,2,...,T}.}

\item{ragged}{A list with number of items equal to length of leads.rag
within the sample {1,2,...,T}.  (ragged=NULL if there are none.)
Each element contains indices of vector components that are missing.
Note: ragged[[j]] must be non-empty, where leads[j] is the time
 where at least one missing value (in-sample) occurs.}

\item{z}{Raw data as N x T matrix with missing values at various time points.
    Missing values are at any 1 <= t <= T, and occur for some of the N series
(the ragged case), and are denoted with a 1i.  That is,
    Im(z[,t]) = rep(1i,N) or subset thereof encodes missing values.}
}
\value{
z: like the input z, but with additional 1i inserted where indicated by ragged
}
\description{
Read multivariate time series and insert 1i for indicated missing values
}
