% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.reg.r
\name{sigex.reg}
\alias{sigex.reg}
\title{Adds regressors to an existing model}
\usage{
sigex.reg(mdl, series, reg)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{series}{Integer between 1 and N, the index of the individual series for
which regressors are being added.}

\item{reg}{A one-column matrix of time series regressors, of length T.
Should have names attribute, as well as start and frequency.}
}
\value{
mdl: the updated sigex model, a list object
}
\description{
Adds regressors to an existing model
}
