% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.renderpd.r
\name{sigex.renderpd}
\alias{sigex.renderpd}
\title{Modifies a covariance matrix so that it is pd}
\usage{
sigex.renderpd(L.mat, D.mat, thresh)
}
\arguments{
\item{L.mat}{Rectangular, lower Cholesky factor}

\item{D.mat}{Vector of diagonal entries in Cholesky decomposition}

\item{thresh}{A threshold for condition numbers.  The new
covariance matrix has a modified D matrix, such
that the condition numbers (with the same L.mat)
are bounded below by thresh.}
}
\value{
conds.new: new condition numbers
}
\description{
Modifies a covariance matrix so that it is pd
}
