% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.resid.r
\name{sigex.resid}
\alias{sigex.resid}
\title{Computes residuals from Gaussian likelihood of model}
\usage{
sigex.resid(psi, mdl, data.ts, debug = FALSE)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}
}
\value{
A list with two elements
	resids: a N x (T-d) matrix of residuals, where d is
		the order of the full differencing polynomial.
		(A failure returns Inf.)
   x.acf: an array of dimension N x (T+1) x N of the autocovariances
     of the differenced model for lags 0 through T, corresponding to psi.
}
\description{
Computes residuals from Gaussian likelihood of model
}
