% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.signal.r
\name{sigex.signal}
\alias{sigex.signal}
\title{Computes signal extraction matrix and error covariance matrix}
\usage{
sigex.signal(data.ts, param, mdl, sigcomps)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal}
}
\value{
list object of f.mat and v.mat
	f.mat: array of dimension c(T,N,T,N), where f.mat[,j,,k]
		is the signal extraction matrix that utilizes input series k
		to generate the signal estimate for series j.
	v.mat:  array of dimension c(T,N,T,N), where v.mat[,j,,k]
		is the error covariance matrix arising from input series k
		used to generate the signal estimate for series j.
}
\description{
Computes signal extraction matrix and error covariance matrix
}
