% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.signalcheck.r
\name{sigex.signalcheck}
\alias{sigex.signalcheck}
\title{Computes model-based signal extraction diagnostics}
\usage{
sigex.signalcheck(signal, param, mdl, sigcomps, lagall)
}
\arguments{
\item{signal}{A T x N matrix ts object of the extracted signal}

\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal}

\item{lagall}{Each signal diagnostic has a lag argument.  They are all
computed, from lags one through lagall}
}
\value{
diagnostics: Complex matrix with lagall rows and N columns, one
		column for each series.  The real part is the diagnostic
		value, given by sample autocorrelation at various lags of
		the estimated signal.  The imaginary part is the t statistic
		of the diagnostic: statistic minus null mean, divided by
		standard error.
}
\description{
Computes model-based signal extraction diagnostics
}
