% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.sim.r
\name{sigex.sim}
\alias{sigex.sim}
\title{Simulate a stochastic process}
\usage{
sigex.sim(psi, mdl, simlen, burnin, dof, init)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{simlen}{Length of the simulation}

\item{burnin}{Initial stretch of simulation, later discarded}

\item{dof}{Innovations of simulation are student t with dof
degrees of freedom, set dof=Inf to get Gaussian}

\item{init}{Initial values for process, should have length d,
where d is the order of the full differencing polynomial}
}
\value{
sims: matrix of dimension simlen x N of simulated stochastic process
		corresponding to mdl with parameter psi.
}
\description{
Simulate a stochastic process
}
