% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.specar.r
\name{sigex.specar}
\alias{sigex.specar}
\title{Plot AR spectrum of input time series}
\usage{
sigex.specar(data.ts, diff = FALSE, subseries, period)
}
\arguments{
\item{data.ts}{A T x N matrix ts object,
corresponding to N time series of length T}

\item{diff}{Boolean, if TRUE difference the series
and plot Growth Rate, else in Levels}

\item{subseries}{Index between 1 and N, indicating which series
to examine}

\item{period}{Number of observations per cycle (e.g. a year or week)}
}
\value{
NA
}
\description{
Plot AR spectrum of input time series
}
