% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.tstats.r
\name{sigex.tstats}
\alias{sigex.tstats}
\title{Computes t statistics for parameter estimates}
\usage{
sigex.tstats(mdl, psi, hess, constraint)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{psi}{A vector of all the real hyper-parameters}

\item{hess}{Hessian matrix, which can be obtained from output of sigex.mlefit}

\item{constraint}{Matrix of the form [Q , C], with C (constraint.mat)
the matrix of constraints and Q (constraint.vec) the vector
of constraint constants, such that C psi = Q.
Use NULL if there are no constraints}
}
\value{
tstats: vector of psi mles divided by standard error.  If a parameter
		constraints are used, this is taken into account.
}
\description{
Computes t statistics for parameter estimates
}
