% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.weekly2daily.r
\name{sigex.weekly2daily}
\alias{sigex.weekly2daily}
\title{De-embeds a weekly time series as a daily time series}
\usage{
sigex.weekly2daily(data.ts, first.day)
}
\arguments{
\item{data.ts}{A T x 7 matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{first.day}{A number between 1 and 7 that indicates what day of the week
should correspond to the first component of the 7-vector.
So 1 = Sunday, 2 = Monday, 3 = Tuesday, 4 = Wednesday,
  5 = Thursday, 6 = Friday, 7 = Sunday.}
}
\value{
data.ts: a T x 1 matrix ts object; any  values to be imputed
		are marked with NA in that entry.  The NA is for missing value.
}
\description{
De-embeds a weekly time series as a daily time series
}
