% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.whichtrend.r
\name{sigex.whichtrend}
\alias{sigex.whichtrend}
\title{Determines which component (index) corresponds to trend}
\usage{
sigex.whichtrend(mdl)
}
\arguments{
\item{mdl}{The specified sigex model, a list object
mdl[[1]] is mdlK, gives ranks of white noise covariance matrix
  mdl[[2]] is mdlType, a list giving t.s. model class, order, and bounds
  mdl[[3]] is mdlDiff, gives delta differencing polynomials
 mdl[[4]] is list of regressors by individual series}
}
\value{
trendcomp: index of the latent component that has a stochastic trend
}
\description{
Determines which component (index) corresponds to trend
}
