% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.whittle.r
\name{sigex.whittle}
\alias{sigex.whittle}
\title{Computes Whittle likelihood of model}
\usage{
sigex.whittle(psi, mdl, data.ts)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any missing values
must be encoded with NA in that entry}
}
\value{
Returns Whittle likelihood, evaluated at Fourier frequencies
		corresponding to model, for differenced time series,
		with hyper-parameter psi.
}
\description{
Computes Whittle likelihood of model
}
