% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.wk.r
\name{sigex.wk}
\alias{sigex.wk}
\title{Computes signal extraction filter coefficients and MSE}
\usage{
sigex.wk(data.ts, param, mdl, sigcomps, target, plotit = TRUE, grid, len)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal S_t}

\item{target}{Array of dimension c(N,N,p+1) for degree p matrix polynomial
phi (B) such that target signal is phi (B) S_t.
normal usage is target <- array(diag(N),c(N,N,1))}

\item{plotit}{Binary flag to indicate whether plots should be generated;
only gives output if N <= 3}

\item{grid}{Desired number of frequencies for spectrum calculations}

\item{len}{Max index of the filter coefficients}
}
\value{
list object of psi.wk and mse.wksig
	psi.wk: array of dimension c(N,N,2*len+1),
		of real number entries, where the third dimension
		indexes coefficients from -len,...,0,...,len
	mse.wksig: N x N matrix of signal extraction mean squared error
}
\description{
Computes signal extraction filter coefficients and MSE
}
