% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.wkextract.r
\name{sigex.wkextract}
\alias{sigex.wkextract}
\title{Computes signal extractions and MSE via WK method}
\usage{
sigex.wkextract(
  psi,
  mdl,
  data.ts,
  sigcomps,
  target,
  grid,
  window,
  horizon,
  needMSE
)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any  values to be imputed
   must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{sigcomps}{Indices of the latent components composing the signal S_t}

\item{target}{Array of dimension c(N,N,p+1) for degree p matrix polynomial
phi (B) such that target signal is phi (B) S_t.
normal usage is target <- array(diag(N),c(N,N,1))}

\item{grid}{Desired number of frequencies for spectrum calculations}

\item{window}{Max index of the WK filter coefficients}

\item{horizon}{A positive integer indicating how many forecasts and
aftcasts of the signal should be generated}

\item{needMSE}{A binary flag, set to 1 if you want MSE based on casting error,
or if there are any missing values; else (with value 0) the routine
runs faster and returns only WK portion of MSE.}
}
\value{
list object of extract.sig, upp, and low
	extract.sig: (T+H) x N matrix of the signal estimates, where H is
		twice the length of horizon
	upp: as extract.sig, plus twice the standard error
	low: as extract.sig, minus twice the standard error
}
\description{
Computes signal extractions and MSE via WK method
}
