% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.wkmse.r
\name{sigex.wkmse}
\alias{sigex.wkmse}
\title{Signal extraction error spectrum from bi-infinite sample}
\usage{
sigex.wkmse(data.ts, param, mdl, sigcomps, grid)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}

\item{sigcomps}{Indices of the latent components composing the signal}

\item{grid}{Desired number of frequencies for spectrum calculations}
}
\value{
frf.wk:  array of dimension c(N,N,grid), with complex number entries
}
\description{
Signal extraction error spectrum from bi-infinite sample
}
