% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.zetalen.r
\name{sigex.zetalen}
\alias{sigex.zetalen}
\title{Computes the length of zeta}
\usage{
sigex.zetalen(mdlType, N)
}
\arguments{
\item{mdlType}{This is a component of mdl (the specified sigex model),
cited as mdl[[2]]}

\item{N}{Cross-section dimension}
}
\value{
zetalen: the length of zeta
}
\description{
Computes the length of zeta
}
