% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specFact.r
\name{specFact}
\alias{specFact}
\title{Compute the spectral factorization of a given p.d. sequence}
\usage{
specFact(poly)
}
\arguments{
\item{poly}{A symmetric vector of coefficients, which is p.d.}
}
\value{
theta: the spectral factorization, such that
poly(z,z^{-1}) = theta(z) * theta(z^{-1})
}
\description{
Compute the spectral factorization of a given p.d. sequence
}
