% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specFactmvar.r
\name{specFactmvar}
\alias{specFactmvar}
\title{Compute the spectral factorization of a given multivariate p.d. sequence}
\usage{
specFactmvar(xAcf)
}
\arguments{
\item{xAcf}{An array containing the given p.d. sequence.  It has
dimension N x (q+1) x N, where xAcf[,1,] corresponds
to the zeroth coefficient matrix, xAcf[,2,] is the first
coefficient matrix, etc.}
}
\value{
theta: the spectral factorization as a list object
		theta[[1]] is an array of dimension N x N x (q+1),
		consisting of the VMA coefficients from index q through index zero
		(Note that theta[[1]][,,q+1] should be an identity matrix.)
		theta[[2]] is the white noise covariance matrix
		theta[[3]] tracks the number of iterations used in this algorithm,
			which is based on Bauer's method, as described in McElroy (2018, JTSA)
}
\description{
Compute the spectral factorization of a given multivariate p.d. sequence
}
