% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubgenerator.r
\name{ubgenerator}
\alias{ubgenerator}
\title{Compute the product of unit root differencing operators.}
\usage{
ubgenerator(period, trunc.len, m, rho)
}
\arguments{
\item{period}{Possibly non-integer period of sinusoids}

\item{trunc.len}{Gives the maximal index of k, and is set to [period/2]
when trunc.len = NULL}

\item{m}{Large integer giving number of cepstral coefficients}
}
\value{
wolds: coefficients of product polynomial.
}
\description{
Compute the product of unit root differencing operators.
}
