% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.par2pre.r
\name{var.par2pre}
\alias{var.par2pre}
\title{generates a stable dimension N VAR(p) process}
\usage{
var.par2pre(param)
}
\arguments{
\item{param}{N x N x var.order array,
where var.order is the VAR order p and N is the dimension of the process}
}
\value{
psi: vector of real numbers, of length p*N^2
}
\description{
generates a stable dimension N VAR(p) process
}
