% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.pre2par.r
\name{var.pre2par}
\alias{var.pre2par}
\title{generates a stable dimension N VAR(p) process}
\usage{
var.pre2par(psi, var.order, N, debug = FALSE)
}
\arguments{
\item{psi}{vector of real numbers, of length p*N^2}

\item{var.order}{the VAR order p}

\item{N}{dimension of the process}

\item{debug}{a Boolean flag; if true, outputs stability check}
}
\value{
param: N x N x var.order array
}
\description{
generates a stable dimension N VAR(p) process
}
