% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var2.par2pre.r
\name{var2.par2pre}
\alias{var2.par2pre}
\title{Generates a stable dimension N VAR(p) process}
\usage{
var2.par2pre(param)
}
\arguments{
\item{param}{A N x N x p array,
where p is the VAR order and N is the dimension of the process}
}
\value{
psi: vector of real numbers, of length p*N^2
}
\description{
Generates a stable dimension N VAR(p) process
}
