% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var2.pre2par.r
\name{var2.pre2par}
\alias{var2.pre2par}
\title{Generates a stable dimension N VAR(p) process}
\usage{
var2.pre2par(psi, p, N)
}
\arguments{
\item{psi}{Vector of real numbers, of length p*N^2}

\item{p}{The VAR order p}

\item{N}{Dimension of the process}
}
\value{
param: N x N x p array
}
\description{
Generates a stable dimension N VAR(p) process
}
