% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VARMA_auto}
\alias{VARMA_auto}
\title{computes autocovariances of VARMA}
\usage{
VARMA_auto(param, p, q, maxlag)
}
\arguments{
\item{p}{- AR order}

\item{q}{- MA order}

\item{maxlag}{- final lag of autocovariance needed}

\item{param:}{matrix of dimension m x (p+q+1)m, equals [ phi | theta | sigma ]
phi: block matrix of dimension m x mp of VAR coefficients
theta: block matrix of dimension m x mq of VMA coefficients
sigma: m x m covariance matrix of white noise}
}
\value{
autocovariances at lags 0 through maxlag, as array of dimension m x m x (maxlag+1)
}
\description{
Background:
   function computes autocovariances of VARMA (p,q) from lag zero
	to maxlag, with array inputs phi and theta.  VARMA equation:
 	(1 - phi[1]B ... - phi[p]B^p) X_t = (1 + theta[1]B ...+ theta[q]B^q) WN_t
}
