% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARMAauto.r
\name{VARMAauto}
\alias{VARMAauto}
\title{Computes autocovariances of VARMA}
\usage{
VARMAauto(phi, theta, sigma, maxlag)
}
\arguments{
\item{phi}{Array of dimension m x m x p of VAR coefficients, e.g.,
phi <- array(cbind(phi1,phi2,...,phip),c(m,m,p))}

\item{theta}{Array of dimension m x m x q of VMA coefficients, e.g.,
theta <- array(cbind(theta1,theta2,...,thetaq),c(m,m,q))}

\item{sigma}{An m x m covariance matrix of white noise}

\item{maxlag}{Final lag of autocovariance needed.}
}
\value{
The autocovariances at lags 0 through maxlag,
   as array of dimension m x m x (maxlag+1)
}
\description{
Computes autocovariances of VARMA
}
