% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x11filters.r
\name{x11filters}
\alias{x11filters}
\title{Generates x11 trend, seasonal, and seasonal adjustment filters.}
\usage{
x11filters(period, p.seas)
}
\arguments{
\item{period}{Possibly non-integer period of seasonality}

\item{p.seas}{Integer number of seasonal moving averages}
}
\value{
List of arrays for trend.filter, seas.filter, and sa.filter
}
\description{
Generates x11 trend, seasonal, and seasonal adjustment filters.
}
