% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymult.r
\name{polymult}
\alias{polymult}
\title{multiply two power series coeficient vectors}
\usage{
polymult(a, b)
}
\arguments{
\item{a}{power series coefficient vector}

\item{b}{power series coefficient vector}
}
\value{
cofficient vector of product
}
\description{
multiply two power series coeficient vectors
}
