% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/polymult.r
\name{polymult}
\alias{polymult}
\title{polymult - Documentation update needed}
\usage{
polymult(a, b)

polymult(a, b)
}
\arguments{
\item{a}{vector of polynomial coefficients, where a[1] is the zeroth coefficient,
a[2] is the first coefficient, etc.}

\item{b}{vector of polynomial coefficients, where b[1] is the zeroth coefficient,
b[2] is the first coefficient, etc.}
}
\value{
product a * b

vector of polynomial coefficients for c(z) = a(z)*b(z),
		where c[1] is the zeroth coefficient, c[2] is the first coefficient, etc.
}
\description{
polymult - Documentation update needed

compute the product of two polynomials
}
\examples{
polymult(c(1, 2), c(1, 4))

}
