% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.add.r
\name{sigex.add}
\alias{sigex.add}
\title{Build the model by adding on another latent component}
\usage{
sigex.add(mdl, vrank, class, order, bounds, name, delta)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{vrank}{Vector of integers between 1 and N, corresponding
to indices of non-zero Schur complements in the GCD
of the innovations' covariance matrix for the new latent component}

\item{class}{Character string of t.s. model type for the new latent component}

\item{order}{Vector of model order}

\item{bounds}{Four numbers, gives bounds for rho and omega,
the cycle parameters of the new latent component
rho lies in (bounds[1],bounds[2])
omega lies in (bounds[3],bounds[4])}

\item{name}{Character string giving the latent component's name}

\item{delta}{Differencing polynomial (corresponds to delta(B) in Background)
written in format c(delta0,delta1,...,deltad)}
}
\value{
mdl: the updated sigex model, a list object
}
\description{
#	Background:
	A sigex model consists of process x = sum y, for
	stochastic components y.  Each component process y_t
	is either stationary or is reduced to stationarity by
	application of a differencing polynomial delta(B), i.e.
		w_t = delta(B) y_t   is stationary.
	We have a model for each w_t process, which is specified
	through the ranks (indices of non-zero Schur complements,
	cf. background for sigex.param2gcd) of the white noise
	covariance matrix; also there is the model type, which
	denotes the specification of the t.s. model for w_t;
	all the regressors, which are specified by individual time series
	rather than by latent component, and must have length T;
	pre-specified bounds for cyclical parameters, for each component,
	if applicable.
}
