% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.adhocextract.r
\name{sigex.adhocextract}
\alias{sigex.adhocextract}
\title{Computes signal extractions and MSE from an ad hoc filter}
\usage{
sigex.adhocextract(psi, mdl, data.ts, adhoc, shift, horizon, needMSE)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any  values to be imputed
   must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{adhoc}{An array N x N x L, where L is length}

\item{shift}{Gives the integer offset for the adhoc filter:
filter coefficients have indices -shift,...,0,...,L-1-shift
set shift = 0 for a causal filter}

\item{horizon}{A non-negative integer indicating how many forecasts and
aftcasts of the signal should be generated}

\item{needMSE}{A binary flag, set to 1 if you want MSE based on casting error,
or if there are any missing values; else (with value 0) the routine
runs faster and returns zero for the MSE.}
}
\value{
list object of extract.sig, upp, and low
	extract.sig: (T+H) x N matrix of the signal estimates, where H is
		twice the length of horizon
	upp: as extract.sig, plus twice the standard error
	low: as extract.sig, minus twice the standard error
}
\description{
Background:
	A sigex model consists of process x = sum y, for
	stochastic components y.  Each component process y_t
	is either stationary or is reduced to stationarity by
	application of a differencing polynomial delta(B), i.e.
		w_t = delta(B) y_t   is stationary.
	We have a model for each w_t process, and can compute its
	autocovariance function (acf), and denote its autocovariance
	generating function (acgf) via gamma_w (B).
	The signal extraction filter for y_t is determined from
	this acgf and delta.
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter
	manifold) together with imaginary component flagging
	whether the hyper-parameter is fixed for purposes of estimation.
}
\details{
Notes:
	method does midcasts for specified time indices,
	applies ad hoc filter for signal (given by adhoc),
	of length L and offset "shift".
	generates signal at all time points t in seq(L-shift-horizon,T-shift+horizon).
}
