% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.blocktoep.r
\name{sigex.blocktoep}
\alias{sigex.blocktoep}
\title{Generates lower triangular array for block Toeplitz matrix}
\usage{
sigex.blocktoep(x.array)
}
\arguments{
\item{x.array}{Array of dimension N x N x H, where x.array[,,1] is gamma(0)}
}
\value{
x.toep: array of dimension N x H x N x H, where x.toep[,j,,k]
}
\description{
Background:
	Given a sequence of matrix gamma(0), gamma(1),...,gamma(k)
	a block Toeplitz matrix has the form
	[ gamma(0), t(gamma(1)),...,t(gamma(k)) ]
	[ gamma(1), gamma(0),...,t(gamma(k-1)) ]
	[ gamma(k),...,gamma(0)]
}
\details{
Notes: presumes that gamma(0) is symmetric
}
