% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.canonize.r
\name{sigex.canonize}
\alias{sigex.canonize}
\title{Compute canonization of a given ARMA process}
\usage{
sigex.canonize(ma.coef, ar.coef)
}
\arguments{
\item{ma.coef}{q coefficients of MA polynomial with unit constant coefficient}

\item{ar.coef}{p coefficients (minus convention) of AR polynomial with
unit constant coefficient}
}
\value{
ma.stab: coefficients of MA polynomial (without a unit constant coefficient),
		 the degree is max(p,q)
}
\description{
Compute canonization of a given ARMA process
}
