% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.conditions.r
\name{sigex.conditions}
\alias{sigex.conditions}
\title{Computes condition number for a covariance matrix}
\usage{
sigex.conditions(data.ts, psi, mdl)
}
\arguments{
\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}

\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}
}
\value{
conds: a S x N matrix of condition numbers, where
		S is the number of components.  Each row gives
		the N condition numbers for the innovation
		covariance matrix of the corresponding latent component.
}
\description{
Background: a non-negative definite matrix Sigma has a
	Generalized Cholesky Decomposition (GCD) of the form
	Sigma = L %*% D %*% t(L),
	where L is unit lower triangular and D is diagonal with
	non-negative entries, referred to as the Schur complements
	of Sigma.  The number of nonzero Schur complements equals
	the rank of Sigma.  The condition numbers can be computed
	by dividing D by the diagonal of Sigma.
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter	manifold)
}
