% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.default.r
\name{sigex.default}
\alias{sigex.default}
\title{Initializes param with zeroes according to the constraints}
\usage{
sigex.default(mdl, data.ts, constraint)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object (with no missing values)
corresponding to N time series of length T}

\item{constraint}{Matrix of the form [Q , C], with C (constraint.mat)
the matrix of constraints and Q (constraint.vec) the vector
of constraint constants, such that C psi = Q.
Use NULL if there are no constraints}
}
\value{
par.default: this is param, filled from a psi of all zeroes,
		subject to constraints.  Will have form specified by mdl.
}
\description{
Initializes param with zeroes according to the constraints
}
