% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.delta.r
\name{sigex.delta}
\alias{sigex.delta}
\title{Compute differencing polynomial with factors omitted}
\usage{
sigex.delta(mdl, omits)
}
\arguments{
\item{mdl}{The specified sigex model, a list object}

\item{omits}{Indices of components that are to be omitted,
when computing the product of differencing polynomials
for the various components.}
}
\value{
Updated differencing polynomial,
	written in format c(delta0,delta1,...,deltad)
}
\description{
Background:
	A sigex model consists of process x = sum y, for
	stochastic components y.  Each component process y_t
	is either stationary or is reduced to stationarity by
	application of a differencing polynomial delta(B), i.e.
		w_t = delta(B) y_t   is stationary.
	The differencing polynomial for x_t is the product of
	the components' polynomials, so long as they are relatively prime
	(this is assumed).  Applying this product polynomial to x_t,
	the effect on a summand y_t is that it is differenced to
	stationary w_t, but a remainder polynomial acts on w_t as well,
	given by the product of all other polynomials.
}
