% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.gausscheck.r
\name{sigex.gausscheck}
\alias{sigex.gausscheck}
\title{Wrapper for Shapiro-Wilks test of normality}
\usage{
sigex.gausscheck(resids)
}
\arguments{
\item{resids}{A T x N matrix of residuals}
}
\value{
tests: output of shapiro.tests routine of R
}
\description{
Wrapper for Shapiro-Wilks test of normality
}
