% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.glr.r
\name{sigex.glr}
\alias{sigex.glr}
\title{Computes the difference of -2*log(Gaussian likelihood) for
	two models, the nested lik minus nesting lik}
\usage{
sigex.glr(data.ts, psi.nested, psi.nesting, mdl.nested, mdl.nesting)
}
\arguments{
\item{data.ts}{A T x N matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{psi.nested}{psi of the nested model}

\item{psi.nesting}{psi of the nesting model}

\item{mdl.nested}{The specified nested sigex model, a list object}

\item{mdl.nesting}{The specified nesting sigex model, a list object}
}
\value{
glr: difference of log likelihoods,
	dof: degrees of freedom, given by different in number of parameters
}
\description{
Background:
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter	manifold)
}
\details{
Format: psi has three portions, psi = [xi,zeta,beta]
	xi ~ all hyper-parameters for covariance matrices
	zeta ~ all hyper-parameters for t.s. models
	beta ~ all regression parameters
Notes: handles missing values in data.ts, which are indicated by 1i.
	This test can be applied to non-nested models,
	but the distribution won't be chi^2
}
