% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.lik.r
\name{sigex.lik}
\alias{sigex.lik}
\title{Computes -2*log(Gaussian likelihood) of model}
\usage{
sigex.lik(psi, mdl, data.ts, debug = TRUE)
}
\arguments{
\item{psi}{A vector of all the real hyper-parameters}

\item{mdl}{The specified sigex model, a list object}

\item{data.ts}{A T x N matrix ts object; any  values to be imputed
    must be encoded with NA in that entry.  The NA is for missing value,
or an enforced imputation (e.g. extreme-value adjustment).}

\item{debug}{Set to TRUE if lik values should be printed to screen}
}
\value{
Value of the divergence
}
\description{
Background:
	param is the name for the model parameters entered into
	a list object with a more intuitive structure, whereas
	psi refers to a vector of real numbers containing all
	hyper-parameters (i.e., reals mapped bijectively to the parameter	manifold)
}
\details{
Format: psi has three portions, psi = [xi,zeta,beta]
	xi ~ all hyper-parameters for covariance matrices
	zeta ~ all hyper-parameters for t.s. models
	beta ~ all regression parameters
Notes: handles missing values in data.ts, which are indicated by 1i
}
