% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.lik.r
\name{sigex.lik}
\alias{sigex.lik}
\title{computes -2*log(Gaussian likelihood) of model}
\usage{
sigex.lik(psi, mdl, data.ts, debug = TRUE)
}
\arguments{
\item{psi}{A  vector of all the real hyper-parameters}

\item{mdl}{the specified sigex model, a list object}

\item{data.ts}{a T x N matrix, corresponding to N time series of length T}

\item{debug}{set to TRUE if lik values should be printed to screen}
}
\value{
value of the divergence
}
\description{
computes -2*log(Gaussian likelihood) of model
}
