% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.lpwk.r
\name{sigex.lpwk}
\alias{sigex.lpwk}
\title{Computes signal extraction filter coefficients
	for trend and cycle, by combining WK filter for trend-cycle
	(specified by trendcyclecomp) with LP filter of cutoff.}
\usage{
sigex.lpwk(data.ts, param, mdl, trendcyclecomp, grid, len, cutoff, trunc)
}
\arguments{
\item{data.ts}{A T x N matrix ts object}

\item{param}{model parameters entered into
a list object with an intuitive structure.}

\item{mdl}{The specified sigex model, a list object}

\item{trendcyclecomp}{The (single) index of the trend-cycle component}

\item{grid}{Desired number of frequencies for spectrum calculations}

\item{len}{Max index of the filter coefficients}

\item{cutoff}{A number between 0 and pi, with all frequencies < cutoff preserved}

\item{trunc}{Truncation index for LP filter}

\item{sigcomps}{Provides indices of a desired component that
is disjoint from trend-cycle, so that MSEs of
trend+sigcomps and cycle+sigcomps are computed.
 (Pass in sigcomps = NULL to just get trend and cycle MSEs.)}
}
\value{
list object with psi.lpwk and psi.ilpwk
	psi.lpwk: array of dimension c(N,N,2*trunc+1), filter coefficients for trend
	psi.ilpwk: array of dimension c(N,N,2*trunc+1), filter coefficients for cycle
}
\description{
Computes signal extraction filter coefficients
	for trend and cycle, by combining WK filter for trend-cycle
	(specified by trendcyclecomp) with LP filter of cutoff.
}
