% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.meaninit.r
\name{sigex.meaninit}
\alias{sigex.meaninit}
\title{Adds trend regressors to an existing model}
\usage{
sigex.meaninit(mdl, data.ts, d)
}
\arguments{
\item{mdl}{The specified sigex model, a list object.
  mdl[[1]] is mdlK, gives ranks of white noise covariance matrix
  mdl[[2]] is mdlType, a list giving t.s. model class, order, and bounds
  mdl[[3]] is mdlDiff, gives delta differencing polynomials
mdl[[4]] is list of regressors by individual series}

\item{data.ts}{A T x N matrix ts object}

\item{d}{Order of time polynomial trend regressor desired, labeled as "Trend".
 However, if a trend component exists in the model,
then this d is ignored, and the number of unit roots
is instead used to determine d.}
}
\value{
mdl: the updated sigex model, a list object
}
\description{
Background:
	x is a multivariate time series (N x T), and each individual series
	can have its distinct set of regressors.  So for each 1 <= j <= N,
	x[j,] is length T and has r_j number of length T regressors.
	There is a default regressor of polynomial time: suppose the
	time series has d unit roots (d >= 0), and this applies to each
	individual series (differencing polynomials are the same for all
	individual series in sigex).  Then the regressor t^d for 1 <= t <= T
	is the default "mean effect".  (Coefficients of lower order time
	polynomial effects cannot be identified.)  When d=0, this is
	just the mean of the process.  (Although it need not be stationary
	when d=0, any other non-stationary latent components are assumed to
	have mean zero for identifiability.)  One can always add higher order
	time polynomial regressors, if desired.
}
\details{
Notes: always use this function when setting up the model.
	First make all calls to sigex.add, then call sigex.meaninit,
	and then add additional regressors (if needed) with sigex.reg.
}
