% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.specar.r
\name{sigex.specar}
\alias{sigex.specar}
\title{plot AR spectrum of input time series}
\usage{
sigex.specar(data.ts, diff = FALSE, subseries)
}
\arguments{
\item{data.ts}{a T x N matrix ts object,
corresponding to N time series of length T}

\item{diff}{boolean, if TRUE difference the series
and plot Growth Rate, else in Levels}

\item{subseries}{index between 1 and N, indicating which series
to examine}
}
\value{
NA
}
\description{
plot AR spectrum of input time series
}
