% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigex.transform.r
\name{sigex.transform}
\alias{sigex.transform}
\title{applies aggregation, followed by transformations to the data}
\usage{
sigex.transform(data.ts, transform, aggregate = FALSE)
}
\arguments{
\item{data.ts}{a T x N matrix ts object,
corresponding to N time series of length T}

\item{transform}{a character indicating an instantaneous
transformation to be applied; current options are
"none", "log", and "logistic"}

\item{aggregate}{a boolean, set to TRUE if all series are to
be aggregated into a total}
}
\value{
data.ts: a T x N0 matrix ts object, where N0=1 if
		aggregate=TRUE, otherwise N0=N
}
\description{
applies aggregation, followed by transformations to the data
}
